+++ examples/03-microservices/src/main.ts
@@ -3,12 +3,16 @@ import { Transport } from '@nestjs/microservices';
 import { ApplicationModule } from './app.module';
 
 async function bootstrap() {
+  const app = await NestFactory.createMicroservice(ApplicationModule);
+  await app.listenAsync();
+
+  /** Hybrid application (HTTP + (n)Microservices)
   const app = await NestFactory.create(ApplicationModule);
   app.connectMicroservice({
     transport: Transport.TCP,
   });
 
   await app.startAllMicroservicesAsync();
-  await app.listen(3001);
+  await app.listen(3001); */
 }
 bootstrap();