+++ gulpfile.js
@@ -59,5 +59,7 @@ gulp.task('move', function() {
     gulp.dest('examples/11-swagger/node_modules/@nestjs')
   ).pipe(
     gulp.dest('examples/12-graphql-apollo/node_modules/@nestjs')
+  ).pipe(
+    gulp.dest('examples/15-mvc/node_modules/@nestjs')
   );
 });