+++ lib/common/interfaces/nest-microservice.interface.d.ts
@@ -3,7 +3,8 @@ import { ExceptionFilter } from './exceptions/exception-filter.interface';
 import { PipeTransform } from './pipe-transform.interface';
 import { NestInterceptor } from './nest-interceptor.interface';
 import { CanActivate } from './can-activate.interface';
-export interface INestMicroservice {
+import { INestApplicationContext } from './nest-application-context.interface';
+export interface INestMicroservice extends INestApplicationContext {
     /**
      * Starts the microservice.
      *
@@ -11,6 +12,12 @@ export interface INestMicroservice {
      * @returns Promise
      */
     listen(callback: () => void): any;
+    /**
+     * Starts the microservice and can be awaited.
+     *
+     * @returns Promise
+     */
+    listenAsync(): Promise<any>;
     /**
      * Setup Web Sockets Adapter, which will be used inside Gateways.
      * Use, when you want to override default `socket.io` library.