+++ lib/core/injector/tokens.d.ts
@@ -1,0 +1,1 @@
+export declare const EXPRESS_REF = "EXPRESS_REF";