+++ src/common/decorators/http/index.ts
@@ -2,3 +2,4 @@ export * from './request-mapping.decorator';
 export * from './route-params.decorator';
 export * from './http-code.decorator';
 export * from './create-route-param-metadata.decorator';
+export * from './render.decorator';