+++ src/core/test/router/router-execution-context.spec.ts
@@ -328,4 +328,31 @@ describe('RouterExecutionContext', () => {
       expect(guardsFn({})).to.eventually.throw();
     });
   });
+  describe('createHandleResponseFn', () => {
+    describe('when "renderTemplate" is defined', () => {
+      it('should call "res.render()" with expected args', () => {
+        const template = 'template';
+        const value = 'test';
+        const response = { render: sinon.spy() };
+        
+        sinon.stub(contextCreator, 'reflectRenderTemplate').returns(template);
+        const handler = contextCreator.createHandleResponseFn(null, true, 100);
+        handler(value, response);
+
+        expect(response.render.calledWith(template, value)).to.be.true;
+      });
+    });
+    describe('when "renderTemplate" is undefined', () => {
+      it('should not call "res.render()"', () => {
+        const result = Promise.resolve('test');
+        const response = { render: sinon.spy() };
+        
+        sinon.stub(contextCreator, 'reflectRenderTemplate').returns(undefined);
+        const handler = contextCreator.createHandleResponseFn(null, true, 100);
+        handler(result, response);
+
+        expect(response.render.called).to.be.false;
+      });
+    });
+  });
 });